refactor(gateway): split config bridge and unify runtime config access
