+++ tests/gateway/test_update_command.py
@@ -642,10 +642,8 @@ class TestUpdateInHelp:
         assert "/update" in result
 
     def test_update_is_known_command(self):
-        """The /update command is in the help text (proxy for _known_commands)."""
-        # _known_commands is local to _handle_message, so we verify by
-        # checking the help output includes it.
-        from gateway.run import GatewayRunner
-        import inspect
-        source = inspect.getsource(GatewayRunner._handle_message)
-        assert '"update"' in source
+        """The /update command is registered as a known gateway command."""
+        from hermes_cli.commands import GATEWAY_KNOWN_COMMANDS, resolve_command
+
+        assert "update" in GATEWAY_KNOWN_COMMANDS
+        assert resolve_command("update").name == "update"
